+++ test/TypeScript/main.ts
@@ -7,7 +7,7 @@
  * @see https://github.com/rochars/byte-data
  */
 
-import * as byteData from '../../main.js'
+import * as byteData from '../../dist/byte-data.umd.js'
 
 let theType = {bits: 16}