+++ test/loader.js
@@ -7,22 +7,10 @@
 let byteData;
 let types = require("binary-data-types");
 
-// Browser bundle
-if (process.argv[3] == '--min') {
-	console.log('min tests');
-    global.window = {};
-    require('../dist/byte-data.min.js');
-    byteData = window.byteData;
-
 // UMD bundle
-} else if (process.argv[3] == '--umd') {
+if (process.argv[3] == '--umd') {
 	console.log('umd tests');
 	byteData = require('../dist/byte-data.umd.js');
-
-// CommonJS dist
-} else if (process.argv[3] == '--cjs') {
-	console.log('cjs tests');
-	byteData = require('../dist/byte-data.cjs.js');
 
 // ES6 dist
 } else if (process.argv[3] == '--esm') {